refactor: move processMilkingCows to its own file
