+++ src/game-logic/reducers/index.test.js
@@ -4,14 +4,12 @@ import {
   FERTILIZERS_PRODUCED,
   LOAN_INCREASED,
   LOAN_PAYOFF,
-  MILKS_PRODUCED,
   PRICE_CRASH,
   PRICE_SURGE,
 } from '../../templates'
 import {
   COW_GESTATION_PERIOD_DAYS,
   COW_HUG_BENEFIT,
-  COW_MILK_RATE_SLOWEST,
   COW_FERTILIZER_PRODUCTION_RATE_SLOWEST,
   MAX_ANIMAL_NAME_LENGTH,
   MAX_DAILY_COW_HUG_BENEFITS,
@@ -29,7 +27,6 @@ import {
   generateCow,
   getCostOfNextStorageExpansion,
   getCowFertilizerItem,
-  getCowMilkItem,
   getCowValue,
   getPlotContentFromItemId,
   getPriceEventForCrop,
@@ -398,108 +395,6 @@ describe('computeStateForNextDay', () => {
     expect(firstRow[0].daysWatered).toBe(1)
     expect(firstRow[0].daysOld).toBe(1)
     expect(todaysNotifications).toBeEmpty()
-  })
-})
-
-describe('processMilkingCows', () => {
-  let state
-
-  beforeEach(() => {
-    state = {
-      cowInventory: [],
-      inventory: [],
-      inventoryLimit: -1,
-      newDayNotifications: [],
-    }
-  })
-
-  describe('cow should not be milked', () => {
-    test('cow is not milked', () => {
-      const baseDaysSinceMilking = 2
-
-      state.cowInventory = [
-        generateCow({
-          daysSinceMilking: baseDaysSinceMilking,
-          gender: genders.FEMALE,
-        }),
-      ]
-
-      const {
-        cowInventory: [{ daysSinceMilking }],
-        inventory,
-        newDayNotifications,
-      } = fn.processMilkingCows(state)
-
-      expect(daysSinceMilking).toEqual(baseDaysSinceMilking)
-      expect(inventory).toEqual([])
-      expect(newDayNotifications).toEqual([])
-    })
-  })
-
-  describe('cow should be milked', () => {
-    describe('inventory space is available', () => {
-      test('cow is milked and milk is added to inventory', () => {
-        state.cowInventory = [
-          generateCow({
-            color: standardCowColors.WHITE,
-            daysSinceMilking: COW_MILK_RATE_SLOWEST,
-            gender: genders.FEMALE,
-          }),
-        ]
-
-        const {
-          cowInventory: [cow],
-          inventory,
-          newDayNotifications,
-        } = fn.processMilkingCows(state)
-
-        const { daysSinceMilking } = cow
-
-        expect(daysSinceMilking).toEqual(0)
-        expect(inventory).toEqual([{ id: 'milk-1', quantity: 1 }])
-        expect(newDayNotifications).toEqual([
-          {
-            message: MILKS_PRODUCED`${{ [getCowMilkItem(cow).name]: 1 }}`,
-            severity: 'success',
-          },
-        ])
-      })
-    })
-
-    describe('inventory space is not available', () => {
-      test('cow is milked but milk is not added to inventory', () => {
-        state.inventoryLimit = 1
-        state.cowInventory = [
-          generateCow({
-            color: standardCowColors.WHITE,
-            daysSinceMilking: COW_MILK_RATE_SLOWEST,
-            gender: genders.FEMALE,
-          }),
-          generateCow({
-            color: standardCowColors.WHITE,
-            daysSinceMilking: COW_MILK_RATE_SLOWEST,
-            gender: genders.FEMALE,
-          }),
-        ]
-
-        const {
-          cowInventory: [cow],
-          inventory,
-          newDayNotifications,
-        } = fn.processMilkingCows(state)
-
-        const { daysSinceMilking } = cow
-
-        expect(daysSinceMilking).toEqual(0)
-        expect(inventory).toEqual([{ id: 'milk-1', quantity: 1 }])
-        expect(newDayNotifications).toEqual([
-          {
-            message: MILKS_PRODUCED`${{ [getCowMilkItem(cow).name]: 1 }}`,
-            severity: 'success',
-          },
-        ])
-      })
-    })
   })
 })
